Convert another batch of fpdfsdk files to avoid shortening
Many of these are just checked_cast, because they are assigned
at API boundaries to types that can't be changed, or used in
arithmetic.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91351
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ fpdfsdk/cpdfsdk_interactiveform.cpp
@@ -155,7 +155,7 @@ CPDFSDK_Widget* CPDFSDK_InteractiveForm::GetWidget(
 void CPDFSDK_InteractiveForm::GetWidgets(
     const WideString& sFieldName,
     std::vector<ObservedPtr<CPDFSDK_Annot>>* widgets) const {
-  for (int i = 0, sz = m_pInteractiveForm->CountFields(sFieldName); i < sz;
+  for (size_t i = 0, sz = m_pInteractiveForm->CountFields(sFieldName); i < sz;
        ++i) {
     CPDF_FormField* pFormField = m_pInteractiveForm->GetField(i, sFieldName);
     DCHECK(pFormField);
